requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-redis==5.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
fcm-django==1.0.12
firebase-admin==5.4.0
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.51.3
grpcio-status==1.51.3
httplib2==0.22.0
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
msgpack==1.0.5
packaging==23.0
pkgutil_resolve_name==1.3.10
proto-plus==1.22.2
protobuf==4.22.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyrsistent==0.19.3
python-decouple==3.8
pytz==2023.2
PyYAML==6.0
redis==4.5.3
requests==2.28.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
zipp==3.15.0
